fix: bind chat turns to native connections
